Cleanup travis build scripts.
before-install.sh and run-script.sh now both given an error if
no PLATFORM was not specified correctly.

Environment variables are now set as part of run-script.sh.  This
makes the travis output simpler and makes it possible to run
run-script.sh outside of travis.

javascript binding generation will not exit early if GH_USER
etc are not set.  This allows the travis scripts to be run in
cloned repos which were otherwise erroring out trying to go
the JS bindings commit.

When generating javascript bindgings, don't username and email
with '--global'.  It should be sufficient to set this only for
the binding repo.