Adds markdown comments in the doxstrings
Since Doxygen 1.8.0 (we are already using it), it is possible to
add markdown comments in the doxygen strings.

This will make our comments easier to read.

This patch  only adds markdown to CCNode.h as a test.

If you like, I'll keep adding markdown comments as I see them.
Please, encourage the rest of the team to do the same.