Move constructor from private to public
In my project I want to add extra functionality to FileUtils. For security reasons I make FileUtils function that can extract encoded data directly to memory without storing it on disk. So I am trying to use FileUtilsAndroid as base class, but it is impossible as constructor is hidden as private member. I offer to move constructor to either public or protected to make it possible to use FileUtilsAndroid as base class.

Another way to solve the problem is to remove constructor as FileUtilsApple does not have a constructor.