Uncomment eaglview release CCEAGLViewImpl-ios
This is leaking the eaglview when the Director is purged.

In our case, we are leaving the cocos environment on some views which
are implemented with cocoa, and because this eaglview is not being
released, it is getting notifications for keyboard events, and trying to
access stuff on the NULL eaglview, crashing our app.

Uncommnenting this line seems to fix the problem for us, and we haven't
been able to find any nasty side effects.