Uniformize image support macro usage.
All images format implementation but WIC test the value of a CC_USE_<FORMAT>
preprocessor macro to verify if the implementation must be available or not.
WIC support is checked against the definition of the macro only,
independently of its value, thus making the implementation available even
if CC_USE_WIC = 0.

This commit makes the tests of the CC_USE_WIC macro identical to the tests
of the other similar image format macros.